nanshe/util/xnumpy.py: Distance transform should not be converted.

+++ nanshe/util/xnumpy.py
@@ -4964,7 +4964,7 @@ def generate_contour(a_image, separation_distance=1.0, margin=1.0):
 
     a_mask_transformed = mahotas.distance(
             a_image
-    ).astype(a_image.dtype)
+    )
 
     above_lower_threshold = (lower_threshold <= a_mask_transformed)
     below_upper_threshold = (a_mask_transformed <= upper_threshold)
